Add files via upload
 With my team, I created a Chrome extension for the grouping of tabs based on a prompt using GPT3. This extension is designed to group various tabs that are opened on chrome having a similar prompt making it easier for the user to group everything related to a specific topic and access it pro-efficiently avoiding a mess. Also, if a new session of chrome is started, the user won't need to search everything once again, he just needs to go into the extensions making it really handy for the users. This project was made with the help of GPT3 and consists of two .js files, one .html file, and one .json file. 
However, there's one problem with this project. Whenever we are trying to upload the project on the chrome extension>manage extensions>developer mode>load unpacked it is showing us an error :
"Manifest is not valid JSON. Line: 6, column: 1, Unexpected token.
Could not load manifest."
After inspecting the cause, we came to know that the manifest.json file is not loaded in chrome.  In the web server an application/json Content-type should be loaded, but instead text/html is being loaded. This is found in inspect>Networks>Header>Content-type. The problem is the browser is not able to load the correct json file and there is no error in the codes of all the files. Did a configuration setting for chrome, but still it shows the same error. We are working on it but eventually, if anyone comes up with the proper solution, we will be glad to fix the problem by trying your solution.